chore(typescript): migrate victory-box-plot to TypeScript
